+++ fabric-1.21.1/src/main/java/de/bene2212/holdmyitemsnf/util/PlatformHooks.java
@@ -1,0 +1,12 @@
+package de.bene2212.holdmyitemsnf.util;
+
+import net.minecraft.world.entity.Entity;
+
+public final class PlatformHooks {
+    private PlatformHooks() {
+    }
+
+    public static boolean isInFluid(Entity entity) {
+        return entity.isInWaterOrBubble() || entity.isInLava();
+    }
+}